Add RedisFailureState
Taken and slightly modified from:
https://github.com/wsargent/circuit_breaker/pull/8

THe modification was to make the `redis` variable a class variable
rather than an instance variable, so that it can be shared amongst
multiple instances of the circuit breaker.

If this code is used for a purpose other than the search circuit
breaker, this change will require more work.