Add additional states, events, and methods
New state: `closed_initial`
The initial, closed state.  Does not have the `reset_failure_count`
call that the original `closed` state has

New event: `manual_reset`
Resets the state to `half_open` from any other state.  Called in
response to following the `enable` link in the UI.

New method: `trip_immediately`
Trips the circuit breaker and sets `failure_count` and
`last_failure_time` to values that will allow the breaker to remain in
an open state for the configured time.  Called in response to following
the `disable` link in the UI.